+++ rusk-test/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 2026-06-10
+
 ### Changed
 
 - Build ledger transactions in test helpers through the explicit canonical
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add rusk test utilities for `dusk-rusk-1.4.4`
 
-[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-rusk-test-1.6.0...HEAD
+[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-rusk-test-1.7.0...HEAD
+[1.7.0]: https://github.com/dusk-network/rusk/compare/dusk-rusk-test-1.6.0...dusk-rusk-test-1.7.0
 [1.6.0]: https://github.com/dusk-network/rusk/compare/dusk-rusk-test-1.4.4-rc.0...dusk-rusk-test-1.6.0
 [1.4.4-rc.0]: https://github.com/dusk-network/rusk/tree/dusk-rusk-test-1.4.4-rc.0